Rename HuggingFace helper module to hf
