+++ neuralset-repo/neuralset/extractors/audio.py
@@ -21,7 +21,7 @@ from neuralset import base as nsbase
 from neuralset.events import etypes
 
 from .base import BaseExtractor
-from .huggingface import HuggingFaceConfig, HuggingFaceMixin
+from .hf import HuggingFaceConfig, HuggingFaceMixin
 
 # pylint: disable=import-outside-toplevel
