+++ neuralset-repo/neuralset/extractors/video.py
@@ -19,7 +19,7 @@ from neuralset import base as nsbase
 from neuralset.events import etypes as evts
 
 from . import base as extractor_base
-from . import huggingface
+from . import hf
 from . import image as image_extractors
 
 logger = logging.getLogger(__name__)
@@ -49,7 +49,7 @@ def resamp_first_dim(data: torch.Tensor, new_first_dim: int) -> torch.Tensor:
     return output
 
 
-class HuggingFaceVideoConfig(huggingface.HuggingFaceConfig):
+class HuggingFaceVideoConfig(hf.HuggingFaceConfig):
     processor_kwargs: dict[str, tp.Any] | None = {"do_rescale": True}
     HF_CLASS_DEFAULTS: tp.ClassVar[dict[str, dict[str, str]]] = {
         "vjepa2": {"processor_cls_name": "AutoVideoProcessor"},
@@ -60,7 +60,7 @@ class HuggingFaceVideoConfig(huggingface.HuggingFaceConfig):
     }
 
 
-class HuggingFaceVideo(extractor_base.BaseExtractor, huggingface.HuggingFaceMixin):
+class HuggingFaceVideo(extractor_base.BaseExtractor, hf.HuggingFaceMixin):
     """Extract video embeddings using a native HuggingFace video model.
 
     Videos are divided into clips of `clip_duration` seconds at the specified
@@ -135,12 +135,12 @@ class HuggingFaceVideo(extractor_base.BaseExtractor, huggingface.HuggingFaceMixi
 
     @classmethod
     def _exclude_from_cls_uid(cls) -> list[str]:
-        return huggingface.HuggingFaceMixin._exclude_from_cls_uid()
+        return hf.HuggingFaceMixin._exclude_from_cls_uid()
 
     def _exclude_from_cache_uid(self) -> list[str]:
         return extractor_base.BaseExtractor._exclude_from_cache_uid(
             self
-        ) + huggingface.HuggingFaceMixin._exclude_from_cache_uid(self)
+        ) + hf.HuggingFaceMixin._exclude_from_cache_uid(self)
 
     def _get_timed_arrays(
         self, events: list[evts.Video], start: float, duration: float
